Make non-core headers protected
JSONModel fails to use JSONModel.h as its umbrella header file.

It has a workaround for building via Cocoapods (JSONModelLib.h), but when building with Carthage or otherwise, Xcode generates erros about public headers not included in the umbrella. The header files not included in JSONModel.h aren't really necessary (and certainly not by CareEv), so this makes them protected.